Pause after assigning Listen in tests to ensure node is handling request
Test flakiness is seen, from time to time, in tests involving Listen
requests. Debugging has shown that the flakiness is due to the node not
having started handling the Listen request when the client is sending
records which it expects to receive back via the Listen. Simply pausing
after assigning the Listen request greatly decreases this flakiness. A real
fix is not possible, without changing the legacy Listen protocol, which
isn't going to happen.