Sort list returned by NodeHashRangesBase.getNodesForHash
Code that is already using this method assumes that the returned list of
node hash ranges is sorted by their `order` field. The code to actually do
this sorting was missing. This commit adds it, along with a unittest to
check the sorting, and updates the code that uses this method.