Always unregister disconnect_detector in ListenRequest
For monitoring the connection state, while waiting for the
StorageEngine, Listen request is using the disconnect_detector. However,
if waiting for the storage engine throws, the disconnect_detector (the
scope class), stays registered in the epoll, causing the trigger and
possible crash, as the stack may not be here anymore.

This patch always unregisters disconnect_detector from epoll, nevermind
the return/exception of the waitEvents.