When DhtTestCase has finished connecting, discard task reference
A task is used to block while establishing a neo connection to the DHT.
Previously, a reference to this task was kept by the test case class and
resumed by the connection notifier, even after connection has finished. To
avoid messing with the test case task unless specifically asked to, the
connection notifier now only resumes the task *while* connection is in
progress. Once connect() has exited, the task reference is discarded.