Change neo GetAll.getNext to work with a delegate, not a buffer copy
This allows the storage engine to pass a slice of its contents to the
request handler's delegate, rather than requiring a copy into the handler's
buffer for each record. (The value retrieved from the storage engine is
immediately copied into the batch buffer, anyway, so the extra copy was
utterly pointless.)

This is an optimisation that reduces CPU load during iteration over large
channels.

Part of https://github.com/sociomantic-tsunami/dhtnode/issues/19