Make all test cases operate on their own channel
With the previous change to the behaviour of the RemoveChannel request, the
Listen tests no longer pass. The reason for this is as follows:
1. There's no way to end a Listen request.
2. This means that whenever a test case starts a Listen request, it
   remains active for the rest of the test suite's life.
3. This means that, after this point, all RemoveChannel requests will fail.
4. RemoveChannel is assigned after every test case, to ensure a clean
   slate for the next test, so 3. breaks all test cases after a Listen
   request has been assigned.

To work around this problem, all tests are configured to use a different
channel. This means that a RemoveChannel request in one test case will
never be blocked by a Listen request of another test case.