Fix flaky Mirror test by only writing a fixed number of records
Previously, the number of records written during the Mirror test could
vary: the Put timer only stopped once a fixed number of Put requests had
been reported to the client as having succeeded. The problem with this
approach is that Put requests may have been handled by the node and
forwarded to the Mirror request before the client is notified that the Put
has succeeded. This made the test flaky, as it was affected by the timing
of the individual requests.

Now, the test always performs a fixed number of Put requests.