Revert "Revert "Merge tag 'neo-v0.11.0' into v14.x.x""
This reverts commit 57003b5c8ddc021ad2c5e09e90153eb8341e279a.

This revert madness is needed to have neo again in neo.

So, v14.6.0 had neo merged in it and then the merge reverted in v14.6.1.
Now merging this release to the neo-v.x.x branch makes neo stop being
neo, because all neo changes are reverted.

No worries, a double revert in git is idempotent, git will just apply an
inverse diff, so what we need to do is revert the revert to get neo back
in neo.

I hope this is clear enough, but if it's not, you can read the whole
story here:
https://mirrors.edge.kernel.org/pub/software/scm/git/docs/howto/revert-a-faulty-merge.txt