Client's NodeHashRanges: Use typeof instead of deprecated Trait
This is deprecated since v5.0.0 but since DMD didn't show deprecation messages on templates,
it was not noticed until we attempted to remove the module.
The replacement that is provided by Ocean is not a 1-to-1 replacement.
However, this code should just be using `int` and `AddrPort` could just define `opCmp`,
which is well defined in the language.
In order to allow releases to move forward, using this single `typeof()` was deemed
to be the least intrusive of all changes.