Update all dependencies to HEAD, fix deprecations, drop compatibility for DMD < 2.092
Since we want to drop support for dmd-transitional,
we need to settle on the minimum supported version of DMD.
Due to the various deprecations and some codeggen bugs in -O -inline,
only v2.092.0 and higher is able to compile our code correctly.

Additionally we now only test on 20.04 instead of xenial,
and reduced coverage to only the latest version of the compiler.

Finally, as Travis is going EOL, we switch to Github Actions,
and ditch Beaver in the process to keep all CI settings in sync.